backend/requirements.txt
django
djangorestframework
python-dotenv
mysqlclient
huggingface-hub 
django-cors-headers
